pyproject.toml
[project]
name = "dit-exam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "nicegui>=3.5.0",
    "openai>=2.15.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
]
